changed to fast12 detector, applied OFfilter, rearranged structure of corner management

+++ modules/OpticFlow/opticflow_code.h
@@ -12,8 +12,8 @@ extern unsigned int verbose;
 extern float opt_trans_x;
 extern float opt_trans_y;
 extern float Velx, Vely;
-extern int opt_angle_x_raw;
-extern int opt_angle_y_raw;
+extern float opt_angle_x_raw;
+extern float opt_angle_y_raw;
 
 extern float divergence;
